Move modal to separate module
